DatasetFiltration/requirements.txt
pandas>=2.0.0
google-generativeai>=0.7.0
tqdm>=4.65.0